Start server in child process
